Split out mock propolis-server from real thing
The simulated sled-agent in Omicron uses the Propolis mock server as a
stand-in for a real VM instance.  In order to break an awkward
dependencies situation, it would be nice if the mock server did not haul
in all everything which propolis-server requires.

Fixes #553
