+++ bin/mock-server/src/main.rs
@@ -1,0 +1,122 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use std::net::SocketAddr;
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use anyhow::anyhow;
+use clap::Parser;
+use dropshot::{ConfigDropshot, HandlerTaskMode, HttpServerStarter};
+use slog::{info, Drain};
+
+#[derive(Debug, Parser)]
+#[clap(about, version)]
+/// An HTTP server providing access to Propolis
+enum Args {
+    /// Generates the OpenAPI specification.
+    OpenApi,
+    /// Runs the Propolis server.
+    Run {
+        #[clap(action)]
+        cfg: PathBuf,
+
+        #[clap(name = "PROPOLIS_IP:PORT", action)]
+        propolis_addr: SocketAddr,
+
+        /// IP:Port for the Oximeter register address
+        #[clap(long, action)]
+        metric_addr: Option<SocketAddr>,
+    },
+}
+
+fn build_logger() -> slog::Logger {
+    let main_drain = if atty::is(atty::Stream::Stdout) {
+        let decorator = slog_term::TermDecorator::new().build();
+        let drain = slog_term::FullFormat::new(decorator).build().fuse();
+        slog_async::Async::new(drain)
+            .overflow_strategy(slog_async::OverflowStrategy::Block)
+            .build_no_guard()
+    } else {
+        let drain =
+            slog_bunyan::with_name("propolis-server", std::io::stdout())
+                .build()
+                .fuse();
+        slog_async::Async::new(drain)
+            .overflow_strategy(slog_async::OverflowStrategy::Block)
+            .build_no_guard()
+    };
+
+    let (dtrace_drain, probe_reg) = slog_dtrace::Dtrace::new();
+
+    let filtered_main = slog::LevelFilter::new(main_drain, slog::Level::Info);
+
+    let log = slog::Logger::root(
+        slog::Duplicate::new(filtered_main.fuse(), dtrace_drain.fuse()).fuse(),
+        slog::o!(),
+    );
+
+    if let slog_dtrace::ProbeRegistration::Failed(err) = probe_reg {
+        slog::error!(&log, "Error registering slog-dtrace probes: {:?}", err);
+    }
+
+    log
+}
+
+pub fn run_openapi() -> Result<(), String> {
+    propolis_mock_server::api()
+        .openapi("Oxide Propolis Server API", "0.0.1")
+        .description(
+            "API for interacting with the Propolis hypervisor frontend.",
+        )
+        .contact_url("https://oxide.computer")
+        .contact_email("api@oxide.computer")
+        .write(&mut std::io::stdout())
+        .map_err(|e| e.to_string())
+}
+
+async fn run_server(
+    config_dropshot: dropshot::ConfigDropshot,
+    _metrics_addr: Option<SocketAddr>,
+    log: slog::Logger,
+) -> anyhow::Result<()> {
+    let context = propolis_mock_server::Context::new(log.new(slog::o!()));
+
+    info!(log, "Starting server...");
+
+    let server = HttpServerStarter::new(
+        &config_dropshot,
+        propolis_mock_server::api(),
+        Arc::new(context),
+        &log,
+    )
+    .map_err(|error| anyhow!("Failed to start server: {}", error))?
+    .start();
+
+    let server_res = server.await;
+    server_res.map_err(|e| anyhow!("Server exited with an error: {}", e))
+}
+
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    // Command line arguments.
+    let args = Args::parse();
+
+    match args {
+        Args::OpenApi => run_openapi()
+            .map_err(|e| anyhow!("Cannot generate OpenAPI spec: {}", e)),
+        Args::Run { cfg: _cfg, propolis_addr, metric_addr } => {
+            // Dropshot configuration.
+            let config_dropshot = ConfigDropshot {
+                bind_address: propolis_addr,
+                request_body_max_bytes: 1024 * 1024, // 1M for ISO bytes
+                default_handler_task_mode: HandlerTaskMode::Detached,
+            };
+
+            let log = build_logger();
+
+            run_server(config_dropshot, metric_addr, log).await
+        }
+    }
+}
